test(appengine): consistently define endpoint types
samplingPeriod was declared as a double in the object table, but the
endpoint was declared as an integer in insert_object_receiving_device_endpoints/1

the typing in the object table was changed to match that of the endpoint,
and also making it consistent with individual value's samplingPeriod

+++ apps/astarte_appengine_api/test/support/database_test_helper.exs
@@ -861,7 +861,7 @@ defmodule Astarte.AppEngine.API.DatabaseTestHelper do
     reception_timestamp timestamp,
     reception_timestamp_submillis smallint,
     v_enable boolean,
-    v_samplingPeriod double,
+    v_samplingPeriod int,
     v_binaryblobarray list<blob>,
     PRIMARY KEY ((device_id, path), reception_timestamp, reception_timestamp_submillis));
     """
